+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- `Promotion Banner` component changes, full clickable and optional text and button ([#23](https://github.com/vtex-sites/accenturebb.store/pull/23))
 - `Brand Section` component to CMS ([#22](https://github.com/vtex-sites/accenturebb.store/pull/22))
 - `Fix layout` ([#24](https://github.com/vtex-sites/accenturebb.store/pull/24))
     - Cart (remove alert "Free Shipping" and translate button "Remove")